Start of using tensors and network with pytorch.

+++ src/models/chorales_dataset.py
@@ -4,10 +4,10 @@ from torch.utils.data import Dataset
 
 class ChoralesDataset(Dataset):
     def __init__(self, sequence_data: dict):
-        self.X_soprano = torch.tensor(sequence_data['soprano'])
-        self.Y_alto = torch.tensor(sequence_data['alto'])
-        self.Y_tenor = torch.tensor(sequence_data['tenor'])
-        self.Y_bass = torch.tensor(sequence_data['bass'])
+        self.X_soprano = torch.tensor(sequence_data['soprano'], dtype=torch.float32)
+        self.Y_alto = torch.tensor(sequence_data['alto'], dtype=torch.float32)
+        self.Y_tenor = torch.tensor(sequence_data['tenor'], dtype=torch.float32)
+        self.Y_bass = torch.tensor(sequence_data['bass'], dtype=torch.float32)
 
         self.length = len(sequence_data['soprano'])
